Fix dangerous casts
On Windows 64-bit, the size of long is 32-bit, thus any pointer to long cast is
not valid. In any case, one should use intptr_t and ptrdiff_t when casting
to integers. This commit attempts to fix all instances reported by GCC.
When relevant, I replaced code by the macros PTR_ADD, ALIGN_UP from system.h

+++ apps/plugins/mikmod/mikmod_internals.h
@@ -703,7 +703,7 @@ extern int Voice_Stopped_internal(SBYTE);
 
 /*========== SIMD mixing helper functions =============*/
 
-#define IS_ALIGNED_16(ptr) (!(((int)(ptr)) & 15))
+#define IS_ALIGNED_16(ptr) (!(((intptr_t)(ptr)) & 15))
 
 /* Altivec helper function */
 #if defined HAVE_ALTIVEC
